# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Groline Backend API",
    "version": "1.0.0",
    "description": "OpenAPI specification generated from the Laravel API routes and controller validation rules. Import this file into Postman and set the server URL or collection base URL for your environment."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Laravel server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum token"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 1
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "nullable": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "Validation error or exception message"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/user": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "GET api/user (user)",
        "operationId": "get_api_user",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/about": {
      "get": {
        "tags": [
          "about"
        ],
        "summary": "GET api/v1/about (about)",
        "operationId": "get_api_v1_about",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/ad/create": {
      "post": {
        "tags": [
          "ad"
        ],
        "summary": "POST api/v1/ad/create (create)",
        "operationId": "post_api_v1_ad_create",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name_ar": {
                    "type": "string"
                  },
                  "name_en": {
                    "type": "string",
                    "nullable": true
                  },
                  "name_fr": {
                    "type": "string",
                    "nullable": true
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "url",
                      "product",
                      "static"
                    ]
                  },
                  "url": {
                    "type": "string",
                    "nullable": true
                  },
                  "product_id": {
                    "type": "integer",
                    "example": 1,
                    "nullable": true
                  }
                },
                "required": [
                  "name_ar",
                  "type"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/ad/delete": {
      "post": {
        "tags": [
          "ad"
        ],
        "summary": "POST api/v1/ad/delete (delete)",
        "operationId": "post_api_v1_ad_delete",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ad_id": {
                    "type": "integer",
                    "example": 1
                  }
                },
                "required": [
                  "ad_id"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ad/restore": {
      "post": {
        "tags": [
          "ad"
        ],
        "summary": "POST api/v1/ad/restore (restore)",
        "operationId": "post_api_v1_ad_restore",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ad_id": {
                    "type": "integer",
                    "example": 1
                  }
                },
                "required": [
                  "ad_id"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/ad/update": {
      "post": {
        "tags": [
          "ad"
        ],
        "summary": "POST api/v1/ad/update (update)",
        "operationId": "post_api_v1_ad_update",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "ad_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "name_ar": {
                    "type": "string"
                  },
                  "name_en": {
                    "type": "string",
                    "nullable": true
                  },
                  "name_fr": {
                    "type": "string",
                    "nullable": true
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "url",
                      "product",
                      "static"
                    ]
                  },
                  "url": {
                    "type": "string",
                    "nullable": true
                  },
                  "product_id": {
                    "type": "integer",
                    "example": 1,
                    "nullable": true
                  }
                },
                "required": [
                  "ad_id"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/brand/create": {
      "post": {
        "tags": [
          "brand"
        ],
        "summary": "POST api/v1/brand/create (create)",
        "operationId": "post_api_v1_brand_create",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name_ar": {
                    "type": "string"
                  },
                  "name_en": {
                    "type": "string",
                    "nullable": true
                  },
                  "name_fr": {
                    "type": "string",
                    "nullable": true
                  },
                  "slug": {
                    "type": "string"
                  },
                  "status": {
                    "type": "boolean"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "name_ar",
                  "slug"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/brand/delete": {
      "post": {
        "tags": [
          "brand"
        ],
        "summary": "POST api/v1/brand/delete (delete)",
        "operationId": "post_api_v1_brand_delete",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "brand_id": {
                    "type": "integer",
                    "example": 1
                  }
                },
                "required": [
                  "brand_id"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/v1/brand/restore": {
      "post": {
        "tags": [
          "brand"
        ],
        "summary": "POST api/v1/brand/restore (restore)",
        "operationId": "post_api_v1_brand_restore",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/brand/update": {
      "post": {
        "tags": [
          "brand"
        ],
        "summary": "POST api/v1/brand/update (update)",
        "operationId": "post_api_v1_brand_update",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "brand_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "name_ar": {
                    "type": "string"
                  },
                  "name_en": {
                    "type": "string",
                    "nullable": true
                  },
                  "name_fr": {
                    "type": "string",
                    "nullable": true
                  },
                  "slug": {
                    "type": "string"
                  },
                  "status": {
                    "type": "boolean"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "brand_id"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/v1/category/delete": {
      "post": {
        "tags": [
          "category"
        ],
        "summary": "POST api/v1/category/delete (delete)",
        "operationId": "post_api_v1_category_delete",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category_id": {
                    "type": "integer",
                    "example": 1
                  }
                },
                "required": [
                  "category_id"
                ]
              }
            }
          }
        }
      }
    },
...
  }
}